Moved Settings styles to styles folder and updated imports
